+++ libs/core-ui/src/lib/Highchart/IHighchartOptions.ts
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { ITheme } from "@fluentui/react";
+
+import { IHighchartsConfig } from "./HighchartTypes";
+
+export interface IHighchartsOptions {
+  configOverride?: IHighchartsConfig;
+  enabled: boolean;
+  theme?: ITheme;
+}
